\name{citations}
\alias{citations}
\title{Plot github stats}
\usage{
  citations(pmid = "25164526", extra = c("yes", "no"))
}
\arguments{
  \item{pmid}{PMID(s) of citations to retrieve}

  \item{extra}{either "yes" or "no"}

}
\description{
  This function fetches citation information from PubMed and formats for use in mediawiki
}
\author{
  Robert Badgett, Scott Chamberlain, Jeroen Ooms
}

